+++ email/src/account/sync.rs
@@ -288,13 +288,13 @@ impl<'a> AccountSyncBuilder {
             FolderSyncStrategy::Include(folders) => FolderSyncStrategy::Include(
                 folders
                     .iter()
-                    .map(|folder| Ok(self.account_config.folder_alias(folder)?))
+                    .map(|folder| Ok(self.account_config.get_folder_alias(folder)?))
                     .collect::<Result<_>>()?,
             ),
             FolderSyncStrategy::Exclude(folders) => FolderSyncStrategy::Exclude(
                 folders
                     .iter()
-                    .map(|folder| Ok(self.account_config.folder_alias(folder)?))
+                    .map(|folder| Ok(self.account_config.get_folder_alias(folder)?))
                     .collect::<Result<_>>()?,
             ),
         };